requirements.txt
# Core Streamlit
streamlit>=1.48.0

# LLM APIs (Multi-model support)
groq>=0.31.0
google-generativeai>=0.8.0
openai>=1.0.0
anthropic>=0.18.0

# LangChain ecosystem
langchain>=0.3.27

# Document processing
PyMuPDF>=1.22.0        # provides 'fitz' for PDF text & rendering
pdfplumber>=0.7.0
python-docx>=0.8.11
docx2txt
requests

# OCR / image processing (recommended for scanned PDFs)
# Note: pytesseract requires the Tesseract binary installed on the host (not just the Python package).
Pillow>=10.0.0
pytesseract>=0.3.10

# Data science and visualization
pandas>=2.0.0
numpy
matplotlib>=3.7.0
seaborn>=0.12.0
plotly

# PDF generation
reportlab>=3.6.0

# Utilities
python-dotenv>=1.1.1
deepdiff

# Optional: cloud OCR (requires credentials)
# google-cloud-vision

# Optional: Vector DB (heavy package, comment out if not used)
# chromadb

# Optional: Gradio (heavy package, comment out if not used)
# gradio>=4.0.0

# Optional: OpenTelemetry (comment out if not used)
# opentelemetry-api==1.37.0
# opentelemetry-exporter-otlp-proto-common==1.37.0
# opentelemetry-proto==1.37.0
# opentelemetry-sdk==1.37.0
# opentelemetry-semantic-conventions==0.58b0
